refactor: ProductBuyBox into smaller components
